+++ src/lib/types/CommonInputProps.ts
@@ -4,7 +4,7 @@ import { FieldValues } from "react-hook-form";
 interface CommonInputProps<T extends FieldValues> {
   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
   onBlur?: (e: React.FocusEvent<HTMLInputElement>) => void;
-  label: ReactNode;
+  label?: ReactNode;
   name: keyof T;
   id?: string;
   helpText?: ReactNode;