+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Labels are now optional for all input fields
+
 ## [0.12.0] - 2023-05-15
 
 ### Added